move load/init plugins to BUILD
`BUILD` is invoked automatically from `->new`, to finish initialising
the object, so those calls should be there: every object requires them!

the trick with declaring the sub and them applying a modifier to it is
needed because we're in a role: if the class that consumes the role
declares its own `BUILD`, the one provided by the role won't be
installed. This way, instead, the empty one won't be installed, but
the modifier will still be applied, and we always get the correct
behaviour